+++ source/gear/buckler-guarding.js
@@ -3,7 +3,7 @@ const { needsLivingTargets } = require('../shared/actionComponents');
 const { addBlock, removeModifier, addModifier } = require('../util/combatantUtil');
 
 module.exports = new GearTemplate("Guarding Buckler",
-	"Grant @{block} block to an ally and yourself and gain @{mod1Stacks} @{mod1}",
+	"Grant an ally @{block} block and gain @{mod1Stacks} @{mod1}",
 	"Block x@{critBonus}",
 	"Armor",
 	"Earth",
@@ -11,20 +11,17 @@ module.exports = new GearTemplate("Guarding Buckler",
 	needsLivingTargets(([target], user, isCrit, adventure) => {
 		let { element, modifiers: [elementStagger, powerUp], block, critBonus } = module.exports;
 		if (user.element === element) {
-			removeModifier(user, elementStagger);
 			removeModifier(target, elementStagger);
 		}
 		if (isCrit) {
 			block *= critBonus;
 		}
 		addBlock(target, block);
-		addBlock(user, block);
 		addModifier(user, powerUp);
-		const userName = user.getName(adventure.room.enemyIdMap);
-		return `Damage will be Blocked for ${target.getName(adventure.room.enemyIdMap)} and ${userName}. ${userName} is Powered Up.`;
+		return `Damage will be Blocked for ${target.getName(adventure.room.enemyIdMap)}. ${user.getName(adventure.room.enemyIdMap)} is Powered Up.`;
 	})
 ).setTargetingTags({ target: "single", team: "delver" })
-	.setSidegrades("Devoted Buckler", "Heavy Buckler")
+	.setSidegrades("Devoted Buckler", "Reinforced Buckler")
 	.setModifiers([{ name: "Stagger", stacks: 1 }, { name: "Power Up", stacks: 25 }])
 	.setDurability(15)
-	.setBlock(75);
+	.setBlock(125);