+++ source/gear/buckler-reinforced.js
@@ -2,8 +2,8 @@ const { GearTemplate } = require('../classes');
 const { needsLivingTargets } = require('../shared/actionComponents');
 const { addBlock, removeModifier, addModifier } = require('../util/combatantUtil');
 
-module.exports = new GearTemplate("Heavy Buckler",
-	"Grant an ally @{block} block and gain @{mod1Stacks} @{mod1}",
+module.exports = new GearTemplate("Reinforced Buckler",
+	"Grant @{block} block to an ally and yourself and gain @{mod1Stacks} @{mod1}",
 	"Block x@{critBonus}",
 	"Armor",
 	"Earth",
@@ -11,17 +11,20 @@ module.exports = new GearTemplate("Heavy Buckler",
 	needsLivingTargets(([target], user, isCrit, adventure) => {
 		let { element, modifiers: [elementStagger, powerUp], block, critBonus } = module.exports;
 		if (user.element === element) {
+			removeModifier(user, elementStagger);
 			removeModifier(target, elementStagger);
 		}
 		if (isCrit) {
 			block *= critBonus;
 		}
 		addBlock(target, block);
+		addBlock(user, block);
 		addModifier(user, powerUp);
-		return `Damage will be Blocked for ${target.getName(adventure.room.enemyIdMap)}. ${user.getName(adventure.room.enemyIdMap)} is Powered Up.`;
+		const userName = user.getName(adventure.room.enemyIdMap);
+		return `Damage will be Blocked for ${target.getName(adventure.room.enemyIdMap)} and ${userName}. ${userName} is Powered Up.`;
 	})
 ).setTargetingTags({ target: "single", team: "delver" })
 	.setSidegrades("Devoted Buckler", "Guarding Buckler")
 	.setModifiers([{ name: "Stagger", stacks: 1 }, { name: "Power Up", stacks: 25 }])
 	.setDurability(15)
-	.setBlock(125);
+	.setBlock(75);