+++ source/gear/inspiration-sweeping.js
@@ -20,7 +20,7 @@ module.exports = new GearTemplate("Sweeping Inspiration",
 		return `Everyone is Powered Up.`;
 	})
 ).setTargetingTags({ target: "all", team: "delver" })
-	.setSidegrades("Reinforcing Inspiration", "Soothing Inspiration")
+	.setSidegrades("Guarding Inspiration", "Soothing Inspiration")
 	.setModifiers([{ name: "Stagger", stacks: 1 }, { name: "Power Up", stacks: 25 }])
 	.setBonus(25) // Power Up stacks
 	.setDurability(10);